requirements.txt
click==8.3.3
numpy==2.4.4
pandas==3.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.33.1
six==1.17.0
urllib3==2.6.3
watchdog==6.0.0
